Fixes 1122705 - Move the ReaderModeButton to its own file
