+++ Client/Frontend/Browser/ReaderModeButton.swift
@@ -1,0 +1,40 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import Foundation
+import UIKit
+
+class ReaderModeButton: UIButton {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setImage(UIImage(named: "reader.png"), forState: UIControlState.Normal)
+        setImage(UIImage(named: "reader_active.png"), forState: UIControlState.Selected)
+    }
+
+    required init(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    var _readerModeState: ReaderModeState = ReaderModeState.Unavailable
+
+    var readerModeState: ReaderModeState {
+        get {
+            return _readerModeState;
+        }
+        set (newReaderModeState) {
+            _readerModeState = newReaderModeState
+            switch _readerModeState {
+            case .Available:
+                self.enabled = true
+                self.selected = false
+            case .Unavailable:
+                self.enabled = false
+                self.selected = false
+            case .Active:
+                self.enabled = true
+                self.selected = true
+            }
+        }
+    }
+}
